+++ src/test/kotlin/io/hhplus/tdd/PointServiceTest.kt
@@ -2,7 +2,7 @@ package io.hhplus.tdd
 
 import io.hhplus.tdd.database.PointHistoryTable
 import io.hhplus.tdd.database.UserPointTable
-import io.hhplus.tdd.point.ErrorCode
+import io.hhplus.tdd.point.exception.ErrorCode
 import io.hhplus.tdd.point.PointService
 import io.hhplus.tdd.point.TransactionType
 import org.assertj.core.api.Assertions.assertThat
